python/requirements.txt
# PS-OCT 处理依赖包
# GPU加速
cupy-cuda12x>=13.0.0  # 根据你的CUDA版本选择, 如cuda11x, cuda12x等

# 核心科学计算
numpy>=1.24.0
scipy>=1.11.0

# 图像处理
opencv-python>=4.8.0
scikit-image>=0.21.0
Pillow>=10.0.0

# 医学图像处理
pydicom>=2.4.0

# 并行计算
joblib>=1.3.0

# 进度条
tqdm>=4.66.0

# 可视化 (可选)
matplotlib>=3.7.0

# MAT文件支持
h5py>=3.9.0
